StreamlitLLMs/BlogPostGeneratorStreamlitLLM/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "BlogPostGeneratorStreamlitLLM"
version = "0.0.1"
description = "A StreamLit App to generate a blog post about a special topic"
authors = [{name = "Mansy", email = "mansy.soroush@gmail.com"}]
dependencies = [
    "ipykernel",
    "python = 3.11.4",
    "python-dotenv",
    "streamlit >= 1.37.0",
    "langchain >= 0.2.11",
    "langchain-openai >= 0.1.19"
]